Backend/requirements.txt
Django==6.0
djangorestframework==3.15.2
django-allauth[google]>=0.63.0
django-cors-headers==4.6.0
django-ratelimit==4.1.0
PyMySQL==1.1.1 # MySQL client for Python
python-decouple==3.8
httpx==0.28.1
requests==2.32.3
PyJWT==2.10.1
cryptography==44.0.1
